requirements.txt
Flask==2.2.5
pymysql==1.0.4
python-dotenv==1.0.0
